requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.11.5        # via requests
chardet==3.0.4            # via requests
idna==2.6                 # via requests
poloniex==0.4.6
python-bitcoinrpc==1.0
requests==2.18.4          # via poloniex
urllib3==1.22             # via requests
